On FreeBSD 12.2 install gets stuck on grep.
Older versions of grep (e.g. grep 2.5.1 on FreeBSD 12.2) read data from
stdin when no file is provided in the command line.

Newer versions (e.g. grep 3.1 on Ubuntu start a search in the working
directory when no file is provided in the command line and the -r flag
is used.

From the grep 2.5.1 manual on FreeBSD 12.2:

"grep searches the named input FILEs (or standard input if no files are
named, or the file name - is given)"

From the grep 3.1 manual on Ubuntu 18.04:

"If no FILE is given, recursive searches examine the working directory,
and nonrecursive searches read standard input."